requirements.txt
requests>=2.31.0
Pillow>=10.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0